refactor: useSignIn
