+++ src/pages-flat/sign-in/sign-in.tsx
@@ -1,13 +1,8 @@
-import { zodResolver } from '@hookform/resolvers/zod'
 import Link from 'next/link'
-import { useRouter } from 'next/router'
-import { useForm } from 'react-hook-form'
-import { toast } from 'react-toastify'
-import { z } from 'zod'
 
 import s from './sign-in.module.scss'
 
-import { useLoginMutation } from '@/features/auth/auth-api'
+import { useSignIn } from '@/pages-flat/sign-in/hooks/useSignIn'
 import { useThirdPartyAuth } from '@/shared/hooks/useThirdPartyAuth'
 import { Button } from '@/shared/ui/button'
 import { Card } from '@/shared/ui/card'
@@ -15,48 +10,9 @@ import { ControlledTextField } from '@/shared/ui/controlled'
 import { GitIcon, GoogleIcon } from '@/shared/ui/icons'
 import { Typography } from '@/shared/ui/typography'
 
-const sigInSchema = z.object({
-  email: z.string().email(),
-  password: z.string().min(3),
-})
-
-type SignInFormShem = z.infer<typeof sigInSchema>
-
 export const SignIn = () => {
   const { onGitHubAuth, onGoogleAuth } = useThirdPartyAuth()
-  const [signIn] = useLoginMutation()
-  const router = useRouter()
-  const {
-    control,
-    handleSubmit,
-    setError,
-    formState: { isValid },
-  } = useForm<SignInFormShem>({
-    defaultValues: {
-      email: '',
-      password: '',
-    },
-    mode: 'onBlur',
-    resolver: zodResolver(sigInSchema),
-  })
-
-  const onSubmit = (data: SignInFormShem) => {
-    signIn(data)
-      .unwrap()
-      .then(res => {
-        localStorage.setItem('access', res.accessToken)
-        router.push('/')
-        toast.success('Success')
-      })
-      .catch(err => {
-        setError('password', {
-          type: 'server',
-          message: err.data.messages,
-        })
-      })
-  }
-
-  const handleSubmitForm = handleSubmit(onSubmit)
+  const { control, handleSubmitForm, isValid } = useSignIn()
 
   return (
     <Card className={s.signBlock}>
